refactor: move tests into a test directory and away from src
