+++ test/send.spec.js
@@ -1,5 +1,5 @@
 import { mocks } from "./index.spec.js";
-import send from "./send.js";
+import send from "../src/send.js";
 
 /**
  * This is a collection of integration tests that make sure modules are doing
